constellation_ai/mvp_cai_2/backend/pyproject.toml
[tool.poetry]
name = "constellation-ai-backend"
version = "0.1.0"
description = "Constellation AI - Internal CRM and knowledge platform"
authors = ["East Rock"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.30.0"
alembic = "^1.13.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.26.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
azure-storage-blob = "^12.19.0"
greenlet = "^3.3.1"
email-validator = "^2.3.0"
pypdf = "^4.0.0"
python-docx = "^1.1.0"
anthropic = "^0.40.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
